.history/requirements_20210203112758.txt
pip==19.2.3
wheel==0.33.6
twine==1.14.0

pandas==0.23.4
biopython==1.73
cipy==1.2.0
tqdm==4.38.0
seaborn==0.11.1
matplotlib==3.3.4 
  
